mbm: sourcers further refactor
